refactor: extract runway segments bar
